+++ apps/repos/views.py
@@ -109,7 +109,6 @@ class RepositorySearchView(ListView):
         context["generated_tag_options"] = repository_json_value_counts("generated_tags")
         params = self.request.GET.copy()
         params.pop("page", None)
-        context["params"] = params
         context["querystring"] = params.urlencode()
         context["total_repositories"] = Repository.objects.count()
         context["total_lists"] = AwesomeList.objects.filter(is_active=True).count()